+++ bin/reset.dart
@@ -1,4 +1,4 @@
-/// Run to reset launcher icons
+/// Run to reset icons launcher
 void main(List<String> arguments) {
-  print('\nReset launcher icons feature not yet implemented. 😭\n');
+  print('\nReset icons launcher feature not yet implemented. 😭\n');
 }